pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "psycopg2>=2.9.12",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "gunicorn>=25.3.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.12",
]
